I have completely restructured the groups url and backend strucutre due to slight complexities to the backend
