+++ src/groups/api/views/groupsView.py
@@ -1,0 +1,7 @@
+from groups.models import Group
+from ..serializers import GroupSerializer
+from rest_framework import viewsets
+
+class GroupsViewSet(viewsets.ModelViewSet):
+    serializer_class = GroupSerializer
+    queryset = Group.objects.all()
